refactor: remove one-file components from folders
