+++ src/components/ui/slide-item.tsx
