+++ src/routes/present/index.tsx
@@ -1,6 +1,6 @@
 import type { DocumentHead } from "@builder.io/qwik-city";
 import { component$ } from "@builder.io/qwik";
-import { PresentationScreen } from "~/components/ui/presentation-screen/presentation-screen";
+import { PresentationScreen } from "~/components/ui/presentation-screen";
 
 export default component$(() => {
   return <PresentationScreen fit="cover" />;
